Place datamodule tests next to datamodules
