+++ project/datamodules/image_classification/fashion_mnist_test.py
@@ -1,0 +1,9 @@
+from project.conftest import setup_with_overrides
+from project.datamodules.image_classification.fashion_mnist import FashionMNISTDataModule
+from project.datamodules.image_classification.image_classification_test import (
+    ImageClassificationDataModuleTests,
+)
+
+
+@setup_with_overrides("datamodule=fashion_mnist")
+class TestFashionMNISTDataModule(ImageClassificationDataModuleTests[FashionMNISTDataModule]): ...
